+++ metadata.rb
@@ -1,11 +1,12 @@
 name 'prometheus'
 maintainer 'Scott M. Likens'
 maintainer_email 'scott@likens.us'
-license "Apache 2.0"
+license 'Apache-2.0'
 description 'Installs/Configures prometheus'
 long_description 'Installs/Configures prometheus'
-version '0.5.0'
+version '1.0.0'
 chef_version '>= 12.14' if respond_to?(:chef_version)
 depends "systemd"
 issues_url 'https://github.com/damm/prometheus/issues'
 source_url 'https://github.com/damm/prometheus'
+supports 'ubuntu'